backend/requirements.txt
Django>=4.2
channels>=4.0
channels-redis>=4.1
celery>=5.3
psycopg2-binary>=2.9
redis>=5.0
flower==2.0.1
flake8>=6.0.0
mypy>=1.0.0
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
django-stubs>=1.16.0
types-redis>=4.5.0
types-requests>=2.31.0
docker>=6.1.3
djangorestframework>=3.14.0 